chore(refactor): refactor folder structure of project - refactoer project structure
