+++ app/src/main/java/com/javadevnairobi/adapter/RecyclerViewAdapter.java
@@ -1,19 +1,18 @@
-package com.javadevnairobi.view;
+package com.javadevnairobi.adapter;
 
 import android.content.Context;
 import android.content.Intent;
 import android.support.v7.widget.CardView;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
 import com.javadevnairobi.R;
+import com.javadevnairobi.view.DetailView;
 
 import java.util.ArrayList;
