+++ app/src/main/java/com/javadevnairobi/adapter/RepoViewAdapter.java
@@ -1,13 +1,10 @@
-package com.javadevnairobi.view;
+package com.javadevnairobi.adapter;
 
 import android.content.Context;
-import android.content.Intent;
-import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
