+++ app/src/main/java/com/javadevnairobi/view/MainActivity.java
@@ -7,12 +7,11 @@ import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.View;
 
 import com.javadevnairobi.R;
+import com.javadevnairobi.adapter.RecyclerViewAdapter;
 import com.javadevnairobi.model.GithubUsers;
 import com.javadevnairobi.presenter.GithubUserPresenter;
 import com.javadevnairobi.utils.NetworkUtility;
